Make the curriculum the main README
It seemed confusing that people had to click 'curriculum' when landing
on the repository home page. Now the 'curriculum' is the main content of
the README.

I moved the content about 'Why Learn C?' into a separate document, and
plan to add to it based on the presentation given by @lifeissweetgood.