Move 'extern variables' to week 4
As I learned more about the use of 'extern' and external variables, it
seems like something rather specific to C and something used in the
context of splitting a project into multiple files. That sort of thing
is our focus in week 4, and would distract from the coverage of
language-agnostic 'basics' in week 2.