Add notes on using multiple files in a project
The notes only cover using header files and splitting your code into
multiple files. There is more to explain related to exactly how the
compiler works, and generating object code, so I'm going to write that
part in a separate set of notes called 'Advanced Compiling'.

The name for that section could probably be improved.