Add notes on standard input and output, file reading and writing
On our original syllabus we had the topic 'sizeof and memory allocation'
for this week, but also had a whole session for 'memory allocation'
later in the syllabus.

Rather than get into the sticky subject of memory allocation when there
are already many other topics for week 4, I think it's worth covering
the use of standard input and output functions, and reading/writing
files. These open up many possibilities for doing small projects, and
are useful for the 'caesar cipher' project, which also is an example of
reusing code between programs.