Move shared values into config for code organization solution
In the solution to the Caesar Cipher problem, we had some 'DEFINE'
constants that were duplicated in several files. Moving these into a
separate file further DRYs up the code.

I'd like to do the same for the alphabet character arrays, *but* when I
do I get a 'duplicate symbol' error on compile.