Add skeletons and solutions for 'Twiddling Bits Part 1'
This adds a skeleton for each of the 'Twiddling Bits' problems, a
'Hints' section with clues for each of the problems in Part 1, and a set
of "solutions".

These all are in need of editing and improvement, but I think that the
skeletons in particular could be helpful for folks starting to go
through the problems.